fix tests for simple policy commissions, move simple policy tests into dedicated file
